[OTAGENT-575] Move regression test to converter_test.go and testdata
Per review: move inline config to testdata/prometheus/unset-env-var/config.yaml
and consolidate the test into converter_test.go alongside the other Convert tests.
